refactor: parser 패키지 추가
- 클래스 조직화를 위해 parser 패키지 추가 및 HttpParser, QueryParser 패키지 이동
